Moved common code to utils
