+++ box_embeddings/common/tf_utils.py
@@ -1,6 +1,17 @@
 import tensorflow as tf
 import math
 import warnings
+from typing import (
+    List,
+    Tuple,
+    Union,
+    Dict,
+    Any,
+    Optional,
+    Type,
+    TypeVar,
+    Callable,
+)
 
 
 def tiny_value_of_dtype(dtype: tf.dtypes.DType) -> float:
@@ -164,3 +175,46 @@ def logsumexp2(t1: tf.Tensor, t2: tf.Tensor) -> tf.Tensor:
 def inv_sigmoid(t1: tf.Tensor) -> tf.Tensor:
     res = tf.math.log(t1 / (1.0 - t1))
     return res
+
+
+def tf_index_select(input_: tf.Tensor, dim: int, indices: List) -> tf.Tensor:
+    """
+    Args:
+        input_(tensor): input tensor
+        dim(int): dimension
+        indices(List): selected indices list
+
+    Returns:
+        Tensor
+    """
+    shape = input_.get_shape().as_list()
+    if dim == -1:
+        dim = len(shape) - 1
+    shape[dim] = 1
+
+    tmp = []
+    for idx in indices:
+        begin = [0] * len(shape)
+        begin[dim] = idx
+        tmp.append(tf.slice(input_, begin, shape))
+    res = tf.concat(tmp, axis=dim)
+
+    return res
+
+
+def _box_shape_ok(t: tf.Tensor) -> bool:
+    if len(t.shape) < 2:
+        return False
+    else:
+        if t.shape[-2] != 2:
+            return False
+
+        return True
+
+
+def _shape_error_str(
+    tensor_name: str, expected_shape: Any, actual_shape: Tuple
+) -> str:
+    return "Shape of {} has to be {} but is {}".format(
+        tensor_name, expected_shape, tuple(actual_shape)
+    )
