+++ box_embeddings/parameterizations/tf_sigmoid_box_tensor.py
@@ -7,36 +7,13 @@ from box_embeddings.parameterizations.tf_box_tensor import (
     TFBoxFactory,
     TFTBoxTensor,
 )
-from box_embeddings.common.tf_utils import softplus_inverse, inv_sigmoid
+from box_embeddings.common.tf_utils import (
+    softplus_inverse,
+    inv_sigmoid,
+    tf_index_select,
+)
 import tensorflow as tf
 import warnings
-
-
-def tf_index_select(
-    input_: tf.Tensor, dim: int, indices: List[int]
-) -> tf.Tensor:
-    """
-    Args:
-        input_(tensor): input tensor
-        dim(int): dimension
-        indices(list): selected indices list
-
-    Returns:
-        Tensor
-    """
-    shape = input_.get_shape().as_list()
-    if dim == -1:
-        dim = len(shape) - 1
-    shape[dim] = 1
-
-    tmp = []
-    for idx in indices:
-        begin = [0] * len(shape)
-        begin[dim] = idx
-        tmp.append(tf.slice(input_, begin, shape))
-    res = tf.concat(tmp, axis=dim)
-
-    return res
 
 
 @TFBoxFactory.register_box_class("tfsigmoid")
